# internal/api/docs/swagger.yaml
basePath: /
definitions:
...
  loads.ErrorBody:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  loads.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/loads.ErrorBody'
    type: object
...
  loads.LoadStats:
    properties:
      equipmentType:
        items:
          $ref: '#/definitions/loads.StatCount'
        type: array
      status:
        items:
          $ref: '#/definitions/loads.StatCount'
        type: array
    type: object
...
  loads.StatCount:
    properties:
      label:
        type: string
      value:
        type: integer
    type: object
  loads.StatsBody:
    properties:
      totals:
        $ref: '#/definitions/loads.LoadStats'
    type: object
  loads.StatsMeta:
    properties:
      numTotal:
        type: integer
      responseTimeMs:
        type: integer
      timestamp:
        type: string
    type: object
  loads.StatsResponse:
    properties:
      meta:
        $ref: '#/definitions/loads.StatsMeta'
      requestURL:
        type: string
      stats:
        $ref: '#/definitions/loads.StatsBody'
    type: object
host: localhost:8080
info:
  contact: {}
  description: |-
    REST API backing the DAT freight demo SPA (AG Grid frontend).
    Implements the AG Grid server-side row model contract
    (pagination, sort, filter, quicksearch) for freight loads,
    plus a single-load lookup endpoint.
  title: dat-freight-demo-go-api
  version: "1.0"
paths:
...
  /loads/stats:
    get:
      description: Returns fixed equipment type and status counts across all loads.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/loads.StatsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/loads.ErrorResponse'
      summary: Get full-dataset freight load statistics
      tags:
      - loads
swagger: "2.0"